renamed cWrapper.h to shtools.h
